Extract components to files. Add move button to nested items.
