+++ frontend/src/pages/character_sheet/panels/InventoryPanel.tsx
@@ -1,40 +1,25 @@
-import CopperCoin from '@assets/images/currency/copper.png';
-import GoldCoin from '@assets/images/currency/gold.png';
-import PlatinumCoin from '@assets/images/currency/platinum.png';
-import SilverCoin from '@assets/images/currency/silver.png';
 import { drawerState } from '@atoms/navAtoms';
-import { EllipsisText } from '@common/EllipsisText';
-import { ItemIcon } from '@common/ItemIcon';
 import { MoveItemMenu } from '@common/operations/item/MoveItemMenu';
 import { isItemVisible } from '@content/content-hidden';
-import { isPlayingStarfinder } from '@content/system-handler';
+
 import classes from '@css/FaqSimple.module.css';
-import { priceToString } from '@items/currency-handler';
 import {
   getBulkLimit,
   getInvBulk,
-  getItemBulk,
-  getItemQuantity,
   handleAddItem,
   handleDeleteItem,
   handleMoveItem,
   handleUpdateItem,
   isItemContainer,
-  isItemEquippable,
   isItemImplantable,
   isItemInvestable,
   isItemWeapon,
-  isItemWithQuantity,
   labelizeBulk,
-  reachedImplantLimit,
-  reachedInvestedLimit,
 } from '@items/inv-utils';
-import { getWeaponStats, parseOtherDamage } from '@items/weapon-handler';
 import {
   Accordion,
   ActionIcon,
   Anchor,
-  Avatar,
   Badge,
   Box,
   Button,
@@ -56,10 +41,11 @@ import { IconCoins, IconMenu2, IconPlus, IconSearch } from '@tabler/icons-react'
 import { ContentPackage, Inventory, InventoryItem, Item, LivingEntity } from '@typing/content';
 import { StoreID } from '@typing/variables';
 import { isPhoneSized } from '@utils/mobile-responsive';
-import { sign } from '@utils/numbers';
 import _ from 'lodash-es';
-import { useState } from 'react';
+import { Key, useState } from 'react';
 import { useRecoilState } from 'recoil';
+import { CurrencySection } from './inventory/CoinsSection';
+import { InvItemOption } from './inventory/InventoryItem';
 
 export default function InventoryPanel(props: {
   id: StoreID;
@@ -105,7 +91,6 @@ export default function InventoryPanel(props: {
       title: (
         <Group wrap='nowrap' gap={20}>
           <Title order={3}>Add Items</Title>
-          {/* (<CurrencySection character={character} />) */}
         </Group>
       ),
       innerProps: {
@@ -261,185 +246,19 @@ export default function InventoryPanel(props: {
             }}
           >
             {_.cloneDeep(invItems)
-              .sort((a, b) => a.item.name.localeCompare(b.item.name))
-              .map((invItem, index) => (
+              .sort((a: InventoryItem, b: InventoryItem) => a.item.name.localeCompare(b.item.name))
+              .map((invItem: InventoryItem, index: number) => (
                 <Box key={index}>
-                  {isItemContainer(invItem.item) ? (
-                    <Accordion.Item className={classes.item} value={`${index}`} w='100%'>
-                      <Accordion.Control>
-                        <Box pr={5}>
-                          <InvItemOption
-                            id={props.id}
-                            entity={props.entity}
-                            isPhone={isPhone}
-                            hideSections
-                            invItem={invItem}
-                            onEquip={(invItem) => {
-                              const newInvItem = _.cloneDeep(invItem);
-                              newInvItem.is_equipped = !newInvItem.is_equipped;
-                              handleUpdateItem(props.setInventory, newInvItem);
-                            }}
-                            onInvest={(invItem) => {
-                              const newInvItem = _.cloneDeep(invItem);
-
-                              if (isItemInvestable(newInvItem.item)) {
-                                newInvItem.is_invested = !newInvItem.is_invested;
-                              }
-                              if (isItemImplantable(newInvItem.item)) {
-                                newInvItem.is_implanted = !newInvItem.is_implanted;
-                              }
-
-                              handleUpdateItem(props.setInventory, newInvItem);
-                            }}
-                            onViewItem={() => {
-                              openDrawer({
-                                type: 'inv-item',
-                                data: {
-                                  storeId: props.id,
-                                  zIndex: 100,
-                                  invItem: _.cloneDeep(invItem),
-                                  onItemUpdate: (newInvItem: InventoryItem) => {
-                                    handleUpdateItem(props.setInventory, newInvItem);
-                                  },
-                                  onItemDelete: (newInvItem: InventoryItem) => {
-                                    handleDeleteItem(props.setInventory, newInvItem);
-                                    openDrawer(null);
-                                  },
-                                  onItemMove: (invItem: InventoryItem, containerItem: InventoryItem | null) => {
-                                    handleMoveItem(props.setInventory, invItem, containerItem);
-                                  },
-                                },
-                                extra: { addToHistory: true },
-                              });
-                            }}
-                          />
-                        </Box>
-                      </Accordion.Control>
-                      <Accordion.Panel>
-                        <Stack gap={5}>
-                          {invItem?.container_contents.map((containedItem, index) => (
-                            <StatButton
-                              key={index}
-                              onClick={() => {
-                                openDrawer({
-                                  type: 'inv-item',
-                                  data: {
-                                    storeId: props.id,
-                                    zIndex: 100,
-                                    invItem: _.cloneDeep(containedItem),
-                                    onItemUpdate: (newInvItem: InventoryItem) => {
-                                      handleUpdateItem(props.setInventory, newInvItem);
-                                    },
-                                    onItemDelete: (newInvItem: InventoryItem) => {
-                                      handleDeleteItem(props.setInventory, newInvItem);
-                                      openDrawer(null);
-                                    },
-                                    onItemMove: (invItem: InventoryItem, containerItem: InventoryItem | null) => {
-                                      handleMoveItem(props.setInventory, invItem, containerItem);
-                                    },
-                                  },
-                                  extra: { addToHistory: true },
-                                });
-                              }}
-                            >
-                              <InvItemOption
-                                id={props.id}
-                                entity={props.entity}
-                                isPhone={isPhone}
-                                invItem={containedItem}
-                                preventEquip
-                                onEquip={(invItem) => {
-                                  const newInvItem = _.cloneDeep(invItem);
-                                  newInvItem.is_equipped = !newInvItem.is_equipped;
-                                  handleUpdateItem(props.setInventory, newInvItem);
-                                }}
-                                onInvest={(invItem) => {
-                                  const newInvItem = _.cloneDeep(invItem);
-
-                                  if (isItemInvestable(newInvItem.item)) {
-                                    newInvItem.is_invested = !newInvItem.is_invested;
-                                  }
-                                  if (isItemImplantable(newInvItem.item)) {
-                                    newInvItem.is_implanted = !newInvItem.is_implanted;
-                                  }
-
-                                  handleUpdateItem(props.setInventory, newInvItem);
-                                }}
-                              />
-                            </StatButton>
-                          ))}
-                          {invItem?.container_contents.length === 0 && (
-                            <Text c='gray.7' fz='sm' ta='center' fs='italic'>
-                              Container is empty
-                            </Text>
-                          )}
-                        </Stack>
-                      </Accordion.Panel>
-                    </Accordion.Item>
-                  ) : (
-                    <Box mb={5}>
-                      <StatButton
-                        key={index}
-                        onClick={() => {
-                          openDrawer({
-                            type: 'inv-item',
-                            data: {
-                              storeId: props.id,
-                              zIndex: 100,
-                              invItem: _.cloneDeep(invItem),
-                              onItemUpdate: (newInvItem: InventoryItem) => {
-                                handleUpdateItem(props.setInventory, newInvItem);
-                              },
-                              onItemDelete: (newInvItem: InventoryItem) => {
-                                handleDeleteItem(props.setInventory, newInvItem);
-                                openDrawer(null);
-                              },
-                              onItemMove: (invItem: InventoryItem, containerItem: InventoryItem | null) => {
-                                handleMoveItem(props.setInventory, invItem, containerItem);
-                              },
-                            },
-                            extra: { addToHistory: true },
-                          });
-                        }}
-                      >
-                        <InvItemOption
-                          id={props.id}
-                          entity={props.entity}
-                          isPhone={isPhone}
-                          invItem={invItem}
-                          onEquip={(invItem) => {
-                            const newInvItem = _.cloneDeep(invItem);
-                            newInvItem.is_equipped = !newInvItem.is_equipped;
-                            handleUpdateItem(props.setInventory, newInvItem);
-                          }}
-                          onInvest={(invItem) => {
-                            const newInvItem = _.cloneDeep(invItem);
-
-                            if (isItemInvestable(newInvItem.item)) {
-                              newInvItem.is_invested = !newInvItem.is_invested;
-                            }
-                            if (isItemImplantable(newInvItem.item)) {
-                              newInvItem.is_implanted = !newInvItem.is_implanted;
-                            }
-
-                            handleUpdateItem(props.setInventory, newInvItem);
-                          }}
-                          additionalButtons={
-                            !isPhone && (
-                              <MoveItemMenu
-                                showOnlyIcon
-                                invItem={invItem}
-                                containerItems={invItems.filter((item) => isItemContainer(item.item))}
-                                onItemMove={(invItem, containerItem) =>
-                                  handleMoveItem(props.setInventory, invItem, containerItem)
-                                }
-                              />
-                            )
-                          }
-                        />
-                      </StatButton>
-                    </Box>
-                  )}
+                  <ItemEntry
+                    invItem={invItem}
+                    index={index}
+                    setInventory={props.setInventory}
+                    id={props.id}
+                    entity={props.entity}
+                    isPhone={isPhone}
+                    openDrawer={openDrawer}
+                    containerItems={invItems.filter((item) => isItemContainer(item.item))}
+                  />
                 </Box>
               ))}
             {invItems.length === 0 && (
@@ -478,225 +297,183 @@ export default function InventoryPanel(props: {
   );
 }
 
-function CurrencySection(props: { entity: LivingEntity | null; onClick?: () => void }) {
-  const pp = props.entity?.inventory?.coins.pp ?? 0;
-  const gp = props.entity?.inventory?.coins.gp ?? 0;
-  const sp = props.entity?.inventory?.coins.sp ?? 0;
-  const cp = props.entity?.inventory?.coins.cp ?? 0;
-
-  const displayAll = isPlayingStarfinder() ? false : true;
-
-  return (
-    <Group
-      gap={15}
-      wrap='nowrap'
-      justify='center'
-      miw={100}
-      style={{
-        cursor: 'pointer',
-      }}
-      onClick={props.onClick}
-    >
-      <CoinSection pp={pp} gp={gp} sp={sp} cp={cp} displayAll={displayAll} justify='center' />
-    </Group>
-  );
-}
-
-export function CoinSection(props: {
-  cp?: number;
-  sp?: number;
-  gp?: number;
-  pp?: number;
-  displayAll?: boolean;
-  justify?: 'flex-start' | 'center' | 'flex-end';
-}) {
-  const pp = props.pp ?? 0;
-  const gp = props.gp ?? 0;
-  const sp = props.sp ?? 0;
-  const cp = props.cp ?? 0;
-  return (
-    <Group gap={15} wrap='nowrap' justify={props.justify}>
-      {(pp || props.displayAll) && (
-        <Group wrap='nowrap' gap={5}>
-          <Text c='gray.4' fz='md' fw={600}>
-            {pp.toLocaleString()}
-          </Text>
-          <Avatar src={PlatinumCoin} alt='Platinum Coins' radius='xs' size='xs' />
-        </Group>
-      )}
-      {(gp || props.displayAll) && (
-        <Group wrap='nowrap' gap={5}>
-          <Text c='gray.4' fz='md' fw={600}>
-            {gp.toLocaleString()}
-          </Text>
-          <Avatar src={GoldCoin} alt='Gold Coins' radius='xs' size='xs' />
-        </Group>
-      )}
-      {(sp || (!pp && !gp && !cp)) && ( // Always show silver coins, even if 0
-        <Group wrap='nowrap' gap={5}>
-          <Text c='gray.4' fz='md' fw={600}>
-            {sp.toLocaleString()}
-          </Text>
-          <Avatar src={SilverCoin} alt='Silver Coins' radius='xs' size='xs' />
-        </Group>
-      )}
-      {(cp || props.displayAll) && (
-        <Group wrap='nowrap' gap={5}>
-          <Text c='gray.4' fz='md' fw={600}>
-            {cp.toLocaleString()}
-          </Text>
-          <Avatar src={CopperCoin} alt='Copper Coins' radius='xs' size='xs' />
-        </Group>
-      )}
-    </Group>
-  );
-}
-
-function InvItemOption(props: {
+function ItemEntry(props: {
   invItem: InventoryItem;
+  index: number;
+  setInventory: React.Dispatch<React.SetStateAction<Inventory>>;
   id: StoreID;
   entity: LivingEntity | null;
-  onEquip?: (invItem: InventoryItem) => void;
-  onInvest?: (invItem: InventoryItem) => void;
-  onViewItem?: (invItem: InventoryItem) => void;
-  hideSections?: boolean;
-  preventEquip?: boolean;
   isPhone?: boolean;
-  additionalButtons?: React.ReactNode;
+  openDrawer: (data: any) => void;
+  containerItems: InventoryItem[];
 }) {
-  const theme = useMantineTheme();
+  const { invItem, openDrawer, containerItems } = props;
 
-  const weaponStats = isItemWeapon(props.invItem.item) ? getWeaponStats(props.id, props.invItem.item) : null;
+  const openDrawerWithItem = (invItem: InventoryItem) => {
+    openDrawer({
+      type: 'inv-item',
+      data: {
+        storeId: props.id,
+        zIndex: 100,
+        invItem: _.cloneDeep(invItem),
+        onItemUpdate: (newInvItem: InventoryItem) => {
+          handleUpdateItem(props.setInventory, newInvItem);
+        },
+        onItemDelete: (newInvItem: InventoryItem) => {
+          handleDeleteItem(props.setInventory, newInvItem);
+          openDrawer(null);
+        },
+        onItemMove: (invItem: InventoryItem, containerItem: InventoryItem | null) => {
+          handleMoveItem(props.setInventory, invItem, containerItem);
+        },
+      },
+      extra: { addToHistory: true },
+    });
+  };
+
+  const equipItem = (invItem: InventoryItem) => {
+    const newInvItem = _.cloneDeep(invItem);
+    newInvItem.is_equipped = !newInvItem.is_equipped;
+    handleUpdateItem(props.setInventory, newInvItem);
+  };
+
+  const investItem = (invItem: InventoryItem) => {
+    const newInvItem = _.cloneDeep(invItem);
+
+    if (isItemInvestable(newInvItem.item)) {
+      newInvItem.is_invested = !newInvItem.is_invested;
+    }
+    if (isItemImplantable(newInvItem.item)) {
+      newInvItem.is_implanted = !newInvItem.is_implanted;
+    }
+
+    handleUpdateItem(props.setInventory, newInvItem);
+  };
+
+  const commonProps = {
+    openDrawerWithItem,
+    equipItem,
+    investItem,
+    otherContainerItems: containerItems.filter((item) => item.id !== invItem.id),
+  } as const;
+
+  if (isItemContainer(invItem.item)) {
+    return <ContainerItemEntry {...props} {...commonProps} />;
+  } else {
+    return <StandardItemEntry {...props} {...commonProps} />;
+  }
+}
+
+function ContainerItemEntry(props: {
+  invItem: InventoryItem;
+  index: number;
+  setInventory: React.Dispatch<React.SetStateAction<Inventory>>;
+  id: StoreID;
+  entity: LivingEntity | null;
+  isPhone?: boolean;
+  otherContainerItems: InventoryItem[];
+  openDrawerWithItem: (invItem: InventoryItem) => void;
+  equipItem: (invItem: InventoryItem) => void;
+  investItem: (invItem: InventoryItem) => void;
+}) {
+  const { invItem, index, isPhone, openDrawerWithItem, equipItem, investItem } = props;
 
   return (
-    <Grid w={'100%'}>
-      <Grid.Col span='auto'>
-        <Group wrap='nowrap' gap={10}>
-          <ItemIcon item={props.invItem.item} size='1.0rem' color={theme.colors.gray[6]} />
-          <Text c='gray.0' fz='sm'>
-            {props.invItem.item.name}
-          </Text>
-          {isItemContainer(props.invItem.item) && props.hideSections && (
-            <Button
-              variant='light'
-              size='compact-xs'
-              radius='xl'
-              onClick={(e) => {
-                e.preventDefault();
-                e.stopPropagation();
-                props.onViewItem?.(props.invItem);
+    <Accordion.Item className={classes.item} value={`${index}`} w='100%'>
+      <Accordion.Control>
+        <Box pr={5}>
+          <InvItemOption
+            id={props.id}
+            entity={props.entity}
+            isPhone={isPhone}
+            hideSections
+            invItem={invItem}
+            onEquip={equipItem}
+            onInvest={investItem}
+            onViewItem={openDrawerWithItem}
+          />
+        </Box>
+      </Accordion.Control>
+      <Accordion.Panel>
+        <Stack gap={5}>
+          {invItem?.container_contents.map((containedItem) => (
+            <StatButton
+              key={containedItem.id}
+              onClick={() => {
+                openDrawerWithItem(containedItem);
               }}
             >
-              View Item
-            </Button>
+              <InvItemOption
+                id={props.id}
+                entity={props.entity}
+                isPhone={isPhone}
+                invItem={containedItem}
+                preventEquip
+                onInvest={investItem}
+                additionalButtons={<DesktopMoveItemMenu {...props} />}
+              />
+            </StatButton>
+          ))}
+          {invItem?.container_contents.length === 0 && (
+            <Text c='gray.7' fz='sm' ta='center' fs='italic'>
+              Container is empty
+            </Text>
           )}
+        </Stack>
+      </Accordion.Panel>
+    </Accordion.Item>
+  );
+}
 
-          {isItemWeapon(props.invItem.item) && weaponStats && (
-            <Group wrap='nowrap' gap={10} maw={300}>
-              <Text c='gray.6' fz='xs' fs='italic' span>
-                {sign(weaponStats.attack_bonus.total[0])}
-              </Text>
-              <EllipsisText c='gray.6' fz='xs' fs='italic' span>
-                {_.truncate(
-                  `${weaponStats.damage.dice}${weaponStats.damage.die}${weaponStats.damage.bonus.total > 0 ? ` + ${weaponStats.damage.bonus.total}` : ``} ${weaponStats.damage.damageType}${parseOtherDamage(weaponStats.damage.other)}${weaponStats.damage.extra ? ` + ${weaponStats.damage.extra}` : ''}`,
-                  { length: props.isPhone ? 15 : 45 }
-                )}
-              </EllipsisText>
-            </Group>
-          )}
-        </Group>
-      </Grid.Col>
-      {!props.isPhone && (
-        <Grid.Col span={3}>
-          <Grid>
-            <Grid.Col span={2}>
-              {!props.hideSections && (
-                <>
-                  {' '}
-                  {isItemWithQuantity(props.invItem.item) && (
-                    <Text ta='center' fz='xs'>
-                      {getItemQuantity(props.invItem.item)}
-                    </Text>
-                  )}
-                </>
-              )}
-            </Grid.Col>
-            <Grid.Col span={3}>
-              {!props.hideSections && (
-                <>
-                  {' '}
-                  <Text ta='center' fz='xs'>
-                    {labelizeBulk(getItemBulk(props.invItem))}
-                  </Text>
-                </>
-              )}
-            </Grid.Col>
-            <Grid.Col span={7}>
-              {!props.hideSections && (
-                <>
-                  {' '}
-                  <Text ta='left' fz='xs'>
-                    {priceToString(props.invItem.item.price)}
-                  </Text>
-                </>
-              )}
-            </Grid.Col>
-          </Grid>
-        </Grid.Col>
-      )}
-      <Grid.Col span={props.isPhone ? 3 : 2} offset={1}>
-        <Group justify='flex-end' wrap='nowrap' align='center' h={'100%'} gap={10}>
-          {isItemInvestable(props.invItem.item) && (
-            <Button
-              size='compact-xs'
-              variant={props.invItem.is_invested ? 'subtle' : 'outline'}
-              color={props.invItem.is_invested ? 'gray.7' : undefined}
-              disabled={!props.invItem.is_invested && reachedInvestedLimit(props.id, props.entity?.inventory)}
-              onClick={(e) => {
-                e.stopPropagation();
-                e.preventDefault();
-                props.onInvest?.(props.invItem);
-              }}
-              w={80}
-            >
-              {props.invItem.is_invested ? 'Divest' : 'Invest'}
-            </Button>
-          )}
-          {isItemImplantable(props.invItem.item) && (
-            <Button
-              size='compact-xs'
-              variant={props.invItem.is_implanted ? 'subtle' : 'outline'}
-              color={props.invItem.is_implanted ? 'gray.7' : undefined}
-              disabled={!props.invItem.is_implanted && reachedImplantLimit(props.id, props.entity?.inventory)}
-              onClick={(e) => {
-                e.stopPropagation();
-                e.preventDefault();
-                props.onInvest?.(props.invItem);
-              }}
-              w={80}
-            >
-              {props.invItem.is_implanted ? 'Extract' : 'Implant'}
-            </Button>
-          )}
-          {isItemEquippable(props.invItem.item) && (
-            <Button
-              size='compact-xs'
-              variant={props.invItem.is_equipped ? 'subtle' : 'outline'}
-              color={props.invItem.is_equipped ? 'gray.7' : undefined}
-              onClick={(e) => {
-                e.stopPropagation();
-                e.preventDefault();
-                props.onEquip?.(props.invItem);
-              }}
-              w={80}
-              disabled={props.preventEquip}
-            >
-              {props.invItem.is_equipped ? 'Unequip' : 'Equip'}
-            </Button>
-          )}
-          {props.additionalButtons}
-        </Group>
-      </Grid.Col>
-    </Grid>
+function StandardItemEntry(props: {
+  invItem: InventoryItem;
+  index: number;
+  setInventory: React.Dispatch<React.SetStateAction<Inventory>>;
+  id: StoreID;
+  entity: LivingEntity | null;
+  isPhone?: boolean;
+  otherContainerItems: InventoryItem[];
+  openDrawerWithItem: (invItem: InventoryItem) => void;
+  equipItem: (invItem: InventoryItem) => void;
+  investItem: (invItem: InventoryItem) => void;
+}) {
+  const { invItem, index, isPhone, openDrawerWithItem, equipItem, investItem } = props;
+  return (
+    <Box mb={5}>
+      <StatButton
+        key={index}
+        onClick={() => {
+          openDrawerWithItem(invItem);
+        }}
+      >
+        <InvItemOption
+          id={props.id}
+          entity={props.entity}
+          isPhone={isPhone}
+          invItem={invItem}
+          onEquip={equipItem}
+          onInvest={investItem}
+          additionalButtons={<DesktopMoveItemMenu {...props} />}
+        />
+      </StatButton>
+    </Box>
+  );
+}
+
+function DesktopMoveItemMenu(props: {
+  isPhone?: boolean;
+  invItem: InventoryItem;
+  otherContainerItems: InventoryItem[];
+  setInventory: React.Dispatch<React.SetStateAction<Inventory>>;
+}) {
+  if (props.isPhone) return null;
+
+  return (
+    <MoveItemMenu
+      showOnlyIcon
+      invItem={props.invItem}
+      containerItems={props.otherContainerItems}
+      onItemMove={(invItem, containerItem) => handleMoveItem(props.setInventory, invItem, containerItem)}
+    />
   );
 }
